Add `git trust` alias for no-bundle-exec solution
Running `git trust` creates an empty `.git/safe` directory, if it doesn't
already exist. The presence of this directory permits the bin/ directory of the
repository to be added to the $PATH. Also runs `hash -r` which tells bash to
discard its executables cache (i.e. force a re-lookup on the new $PATH).