Do not link .rubocop.yml into home directory
RuboCop automatically inherits config from .rubocop.yml files that exist
in ancestor directories. This means that if we have a ~/.rubocop.yml
file, it can affect how rubocop works in other projects, even if those
projects have their own .rubocop.yml files. Not cool.

It is still helpful to have a .rubocop.yml file in this repo, because we
have some ruby scripts in bin/ that could use some linting. But we no
longer will link it into ~.