Simplify `r` script
In modern versions of rails, `bin/rails` will always be present in the
project. Also `bin/rails` will automatically execute rake tasks, so it
can basically be used for everything. The exception is if we want to
call `rake` with no arguments (i.e. the default rake task). So if no
arguments are given, fall back to using `rake` as before.

Of course, if it is a non-rails project, always use `rake` regardless of
what arguments are given.